requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.35
langchain-google-genai
python-dotenv==1.0.1
streamlit==1.37.1
faiss-cpu
pypdf
beautifulsoup4